[react-interactions] Repurpose React a11y modules
